+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 #### Changed
 
-- []() Remove ActiveRecord::Relation#calculate patch
+- [#1199](https://github.com/rails-sqlserver/activerecord-sqlserver-adapter/pull/1199) Remove ActiveRecord::Relation#calculate patch
 
 ## v7.1.3