+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 #### Changed
 
-- []() Remove ActiveRecord::Relation#calculate patch
+- [#1200](https://github.com/rails-sqlserver/activerecord-sqlserver-adapter/pull/1200) Remove ActiveRecord::Relation#calculate patch
 
 ## v7.0.6