+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+#### Added
+
+- [#1201](https://github.com/rails-sqlserver/activerecord-sqlserver-adapter/pull/1201) Support non-dbo schemas in schema dumper.
+
 #### Changed
 
 - [#1153](https://github.com/rails-sqlserver/activerecord-sqlserver-adapter/pull/1153) Only support Ruby v3.1+